requirements.txt
altgraph==0.17.4
iniconfig==2.1.0
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyMsgBox==2.0.1
pyperclip==1.10.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
pytest-mock==3.15.1
python3-xlib==0.15
pytweening==1.2.0
